Create get_F5RealTimeDat
I wanted to share you a little tip to get RealTime information from F5, well it is not the panacea but can help to have more information or may be to do a monitor very important to production environments. Was difficult to implement it, that'cos the administrator should share us the admin credentials, I mean just the admin user can implement this.